+++ .gitignore
@@ -15,6 +15,7 @@
 .vscode/
 .idea/
 go.work
+go.work.sum
 
 # dist packages
 dist/
